+++ Samples~/Showcase/Scripts/HudGuide.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace Hapbeat.Samples.Tutorial
+namespace Hapbeat.Samples.Showcase
 {
     /// <summary>
     /// On-screen help text. ZoneSwitcher と連動して、現在ゾーンのコマンド