+++ Samples~/Showcase/Scripts/SimpleFPSController.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace Hapbeat.Samples.Tutorial
+namespace Hapbeat.Samples.Showcase
 {
     /// <summary>
     /// Minimal keyboard / mouse FPS controller for the Tutorial sample.