+++ Samples~/Showcase/Scripts/SoundPlayer.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Hapbeat.Samples.Tutorial
+namespace Hapbeat.Samples.Showcase
 {
     /// <summary>
     /// Named AudioClip を Inspector で登録し、Animation Event / UnityEvent /