+++ Samples~/Showcase/Scripts/StreamDemoController.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
 
-namespace Hapbeat.Samples.Tutorial
+namespace Hapbeat.Samples.Showcase
 {
     /// <summary>
     /// Z4 Stream Console: <see cref="HapbeatUnityEventTrigger"/> (entry: stream_demo,