+++ Samples~/Showcase/Scripts/TargetReceiver.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace Hapbeat.Samples.Tutorial
+namespace Hapbeat.Samples.Showcase
 {
     /// <summary>
     /// Z5 Target Range hit detector. Projectile の tag に応じて