+++ Samples~/Showcase/Scripts/UiDeselectOnPointerUp.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.EventSystems;
 
-namespace Hapbeat.Samples.Tutorial
+namespace Hapbeat.Samples.Showcase
 {
     /// <summary>
     /// UI 要素 (Slider 等) を pointer up (= マウスドラッグ離した) 時点で